+++ tests/lunes.Domain.UnitTests/Accounts/WhenAddingARevenue.cs
@@ -44,7 +44,7 @@ namespace lunes.Domain.UnitTests.Accounts
 	    [InlineData(300, 260, 560)]
 	    [InlineData(3000, 250.86, 3250.86)]
 	    [InlineData(-2000, 4020, 2020)]
-	    public void ShouldCalculateBalanceCorrectly(double initialBalance, double expense, double expectedBalance)
+	    public void ShouldCalculateBalanceCorrectly(decimal initialBalance, decimal expense, decimal expectedBalance)
 	    {
 			var sut = CreateAccount("Sut Account");
